fix: [tagGroup] Move the inteface declaration to inteface.ts
